+++ lib/common/consts.ex
@@ -9,7 +9,7 @@ defmodule HumanTime.Common.Consts do
   
   @relative_name "(today|tomorrow|yesterday)"
   @relative_adjuster "(this|next|last|week this|week next|week)"
-  @amounts "([0-9]+|one|two|three|four|five|six|seven|eight|nine|ten)"
+  @amounts "([0-9]+|one|two|three|four|five|six|seven|eight|nine|ten|\\ba\\b|\\ban\\b)"
   @iterators "every|every other"
   
   @sep "[-/]"
@@ -23,7 +23,6 @@ defmodule HumanTime.Common.Consts do
   @time_24h "(?P<hour24>[01]?[0-9]|2[0-3]):?(?P<minute24>[0-5][0-9])(:(?P<second>[0-5][0-9]))?"
   @time_all "(?:#{@time_12h}|#{@time_24h}|#{@time_term}|#{@time_current})"
 
-  @amount "[0-9]+"
   @period "(second|minute|hour|day|week|month|year)"
 
   @time_term_compiled @time_term |> Regex.compile!
@@ -59,7 +58,6 @@ defmodule HumanTime.Common.Consts do
     |> String.replace("#AMOUNT#", @amounts)
     |> String.replace("#RELATIVE_NAME#", @relative_name)
     |> String.replace("#RELATIVE_ADJUSTER#", @relative_adjuster)
-    |> String.replace("#AMOUNT#", @amount)
     |> String.replace("#PERIOD#", @period)
     |> Regex.compile!
   end