+++ lib/common/string_lib.ex
@@ -18,6 +18,8 @@ defmodule HumanTime.Common.StringLib do
   # Used to convert from a string into a numrical value
   @doc false
   @spec convert_amount(String.t) :: Integer
+  def convert_amount("a"), do: 1
+  def convert_amount("an"), do: 1
   def convert_amount("one"), do: 1
   def convert_amount("two"), do: 2
   def convert_amount("three"), do: 3